renames scicrunch_api -> scicrunch.service_client
