+++ services/web/server/src/simcore_service_webserver/groups_classifiers.py
@@ -18,8 +18,8 @@ from pydantic import BaseModel, Field, HttpUrl, ValidationError, constr, validat
 from simcore_postgres_database.models.classifiers import group_classifiers
 
 from .constants import APP_DB_ENGINE_KEY
-from .scicrunch.scicrunch_api import SciCrunchAPI
 from .scicrunch.scicrunch_db import ResearchResourceRepository
+from .scicrunch.service_client import SciCrunchAPI
 
 logger = logging.getLogger(__name__)
