+++ services/web/server/src/simcore_service_webserver/groups_handlers.py
@@ -14,9 +14,9 @@ from .groups_exceptions import (
     UserInsufficientRightsError,
 )
 from .login.decorators import RQT_USERID_KEY, login_required
-from .scicrunch.scicrunch_api import SciCrunchAPI
 from .scicrunch.scicrunch_db import ResearchResourceRepository
 from .scicrunch.scicrunch_models import ListOfResourceHits, ResearchResource
+from .scicrunch.service_client import SciCrunchAPI
 from .security_decorators import permission_required
 from .users_exceptions import UserNotFoundError
