+++ services/web/server/src/simcore_service_webserver/scicrunch/__init__.py
@@ -1,6 +1,6 @@
 """
    Submodule to interact with K-Core's https://scicrunch.org service
-    - client to validate and get info about RRIDs via scicrunch's API (scicrunch_api)
+    - client to validate and get info about RRIDs via scicrunch's API (service_client)
     - keeps validated RRIDs in pg-database (scicrunch_db)
     - define models for all interfaces: scicrunch API, postgres DB and webserver API (scicrunch_models)
